update checkstyle warnings fix.
